+++ lib/nostr_sdk.dart
@@ -7,4 +7,5 @@ export 'src/rust/api/protocol/event/tag.dart';
 export 'src/rust/api/protocol/key.dart';
 export 'src/rust/api/protocol/key/public_key.dart';
 export 'src/rust/api/protocol/key/secret_key.dart';
+export 'src/rust/api/relay/options.dart';
 export 'src/rust/api/client.dart';